backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.7
google-auth==2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client==2.116.0
python-dotenv==1.0.1
supabase==2.0.3
groq>=0.9.0
pydantic==2.6.1
requests==2.31.0
pypdf==4.0.1
langchain-text-splitters==0.2.0
google-genai
slowapi==0.1.9
cachetools==5.3.2
